order seeder registration in database seeder file when relation exists

+++ src/Console/RollbackCrudCommand.php
@@ -178,7 +178,7 @@ class RollbackCrudCommand extends BaseCommand
         $this->filesystem->delete($path);
 
         foreach ($this->fields as $field) {
-            if ($this->isSimpleManyToManyRelation($field)) {
+            if (is_array($field) && $this->isSimpleManyToManyRelation($field)) {
                 $migration_name = Str::snake($this->getIntermediateClassName($field));
                 $path = Arr::first($this->filesystem->glob($migrations_path . "/*{$migration_name}"));
                 $this->filesystem->delete($path);
